src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadowdom-dynamic-styling-expected.txt

CONSOLE ERROR: line 15: Uncaught TypeError: box.createShadowRoot is not a function
Test for https://bugs.webkit.org/show_bug.cgi?id=92899. Dynamically styling ShadowDom content on a node distributed to another shadow insertion point fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Content required here to reproduce bug.
